Ask notification permission on SDK>=33 close #2143
