+++ app/src/main/java/com/money/manager/ex/tutorial/TutorialActivity.java
@@ -68,7 +68,6 @@ public class TutorialActivity extends FragmentActivity {
         skipTextView = findViewById(R.id.skipTextView);
         skipTextView.setOnClickListener(view -> onCloseClicked());
 
-        checkNotificationChannel();
     }
 
     @Override
@@ -122,30 +121,6 @@ public class TutorialActivity extends FragmentActivity {
 
         // close
         finish();
-    }
-
-
-    private void checkNotificationChannel() {
-        if (NotificationManagerCompat.from(getApplicationContext()).areNotificationsEnabled()) {
-            return;
-        }
-        
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU)
-        {
-            if (!(checkPermissionGranted(Manifest.permission.POST_NOTIFICATIONS) ) ) {
-                requestPostNotificationsPermission();
-            }
-        }
-    }
-
-    private boolean checkPermissionGranted(String permissions)
-    {
-        // Check if the permission is already available.
-        return (ActivityCompat.checkSelfPermission(this, permissions) == PackageManager.PERMISSION_GRANTED);
-    }
-
-    private void requestPostNotificationsPermission() {
-            ActivityCompat.requestPermissions(this, new String[]{Manifest.permission.POST_NOTIFICATIONS}, 1);
     }
 
 }
